Move exceptions into zeroconf.exceptions
