+++ zeroconf/exceptions.py
@@ -1,0 +1,51 @@
+""" Multicast DNS Service Discovery for Python, v0.14-wmcbrine
+    Copyright 2003 Paul Scott-Murphy, 2014 William McBrine
+
+    This module provides a framework for the use of DNS Service Discovery
+    using IP multicast.
+
+    This library is free software; you can redistribute it and/or
+    modify it under the terms of the GNU Lesser General Public
+    License as published by the Free Software Foundation; either
+    version 2.1 of the License, or (at your option) any later version.
+
+    This library is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+    Lesser General Public License for more details.
+
+    You should have received a copy of the GNU Lesser General Public
+    License along with this library; if not, write to the Free Software
+    Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
+    USA
+"""
+
+# Exceptions
+
+
+class Error(Exception):
+    pass
+
+
+class IncomingDecodeError(Error):
+    pass
+
+
+class NonUniqueNameException(Error):
+    pass
+
+
+class NamePartTooLongException(Error):
+    pass
+
+
+class AbstractMethodException(Error):
+    pass
+
+
+class BadTypeInNameException(Error):
+    pass
+
+
+class ServiceNameAlreadyRegistered(Error):
+    pass
